last

+++ src/Pages/Dashboard/AllUsers/IndividualUser/IndividualUser.js
@@ -28,13 +28,18 @@ const IndividualUser = ({ id, no, name, user_email, data }) => {
    }
 
 
-  const handleDelete = (id) => {
-    fetch(`https://eece-server.vercel.app/allRegisterUsers/${id}`, {
-        method: 'DELETE',
+   const handleDelete = (id) => {
+    console.log(" deleted id",id);
+    if (id) {
+      fetch(`https://eece-server.vercel.app/allRegisterUsers/${id}`, {
+        method: "DELETE",
       })
-      .then(res => res.json())
-      .then(data => console.log("delete user", data))
-  }
+        .then((res) => res.json())
+        .then((data) => console.log(data));
+    }
+  };
+
+
     return (
         <div className="card bg-base-100 shadow-xl">
             <div className="card-body">
@@ -42,7 +47,7 @@ const IndividualUser = ({ id, no, name, user_email, data }) => {
                 <p>{user_email}</p>
                 <div className="card-actions justify-end">
                     { data?.role !== 'admin' && <button onClick={ () => handleMakeAdmin(id) } className="btn bg-custom-pink hover:bg-hover-bg btn-sm">Make admin</button>}
-                    <button onClick={() => handleDelete(id)}  className="btn bg-red-600 hover:bg-red-900 btn-sm">Delete</button>
+                    <button onClick={ () => handleDelete(id)}  className="btn bg-red-600 hover:bg-red-900 btn-sm">Delete</button>
                 </div>
             </div>
         </div>
